+++ tests/api/test_security.py
@@ -1,61 +1,35 @@
-"""Test API security layer."""
+"""Test Supervisor API."""
 
-from aiohttp import web
 import pytest
 
-from supervisor.api import RestAPI
-from supervisor.const import CoreState
 from supervisor.coresys import CoreSys
 
-# pylint: disable=redefined-outer-name
 
+@pytest.mark.asyncio
+async def test_api_security_options_force_security(api_client, coresys: CoreSys):
+    """Test security options force security."""
+    assert not coresys.security.force
 
-@pytest.fixture
-async def api_system(aiohttp_client, run_dir, coresys: CoreSys):
-    """Fixture for RestAPI client."""
-    api = RestAPI(coresys)
-    api.webapp = web.Application()
-    await api.load()
+    await api_client.post("/security/options", json={"force_security": True})
 
-    api.webapp.middlewares.append(api.security.system_validation)
-    yield await aiohttp_client(api.webapp)
+    assert coresys.security.force
 
 
 @pytest.mark.asyncio
-async def test_api_security_system_initialize(api_system, coresys: CoreSys):
-    """Test security."""
-    coresys.core.state = CoreState.INITIALIZE
+async def test_api_security_options_content_trust(api_client, coresys: CoreSys):
+    """Test security options content trust."""
+    assert coresys.security.content_trust
 
-    resp = await api_system.get("/supervisor/ping")
-    result = await resp.json()
-    assert resp.status == 400
-    assert result["result"] == "error"
+    await api_client.post("/security/options", json={"content_trust": False})
 
-
-@pytest.mark.asyncio
-async def test_api_security_system_setup(api_system, coresys: CoreSys):
-    """Test security."""
-    coresys.core.state = CoreState.SETUP
-
-    resp = await api_system.get("/supervisor/ping")
-    result = await resp.json()
-    assert resp.status == 400
-    assert result["result"] == "error"
+    assert not coresys.security.content_trust
 
 
 @pytest.mark.asyncio
-async def test_api_security_system_running(api_system, coresys: CoreSys):
-    """Test security."""
-    coresys.core.state = CoreState.RUNNING
+async def test_api_security_options_pwned(api_client, coresys: CoreSys):
+    """Test security options pwned."""
+    assert coresys.security.pwned
 
-    resp = await api_system.get("/supervisor/ping")
-    assert resp.status == 200
+    await api_client.post("/security/options", json={"pwned": False})
 
-
-@pytest.mark.asyncio
-async def test_api_security_system_startup(api_system, coresys: CoreSys):
-    """Test security."""
-    coresys.core.state = CoreState.STARTUP
-
-    resp = await api_system.get("/supervisor/ping")
-    assert resp.status == 200
+    assert not coresys.security.pwned
